backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.13
